+++ .tape.js
@@ -1,0 +1,7 @@
+module.exports = {
+	'postcss-space-start-attrs': {
+		'basic': {
+			message: 'supports basic usage'
+		}
+	}
+};