+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 1.0.0 (2015-11-16)
+# Changes to Space Start Attributes
+
+### 2.0.0 (December 6, 2016)
+
+- Updated: boilerplate conventions (Node v6.9.1 LTS)
+
+### 1.0.1 (December 7, 2015)
+
+- Updated: Documentation
+
+### 1.0.0 (November 16, 2015)
 
 - Added: Initial version